requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.2=py310haa95532_0
autopep8=1.6.0=pyhd3eb1b0_1
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
certifi=2022.12.7=py310haa95532_0
django=4.1=py310haa95532_0
django-cors-headers=3.14.0=pypi_0
djangorestframework=3.14.0=pypi_0
djangorestframework-gis=1.0=pypi_0
gdal=3.4.3=pypi_0
krb5=1.19.4=h5b6d351_0
libffi=3.4.2=hd77b12b_6
libpq=12.9=hb652d5d_3
openssl=1.1.1t=h2bbff1b_0
pillow=9.4.0=pypi_0
pip=23.0.1=py310haa95532_0
psycopg2=2.9.3=py310hcd4344a_0
pycodestyle=2.10.0=py310haa95532_0
python=3.10.9=h966fe2a_1
python-tzdata=2021.1=pyhd3eb1b0_0
pytz=2022.7.1=pypi_0
setuptools=65.6.3=py310haa95532_0
sqlite=3.40.1=h2bbff1b_0
sqlparse=0.4.3=py310haa95532_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tzdata=2022g=h04d1e81_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wheel=0.38.4=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
zlib=1.2.13=h8cc25b3_0
